requirements.txt
asgiref==3.8.1
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==0.63.6
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.2
factory-boy==3.3.0
Faker==26.0.0
idna==3.7
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
